+++ src/main/java/com/memora/server/entity/HealthReport.java
@@ -16,7 +16,7 @@ public class HealthReport {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long reportId;
+    private Integer reportId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
