+++ src/main/java/com/memora/server/repository/DiarySegmentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DiarySegmentRepository extends JpaRepository<DiarySegment, Long> {
+public interface DiarySegmentRepository extends JpaRepository<DiarySegment, Integer> {
     // 특정 일기에 속한 세그먼트들을 순서(stepOrder)대로 조회
     List<DiarySegment> findByDiaryOrderByStepOrderAsc(Diary diary);
 }
